refactor(console): Config & Security
