+++ console/src/pages/Agent/Config/components/index.ts
@@ -1,0 +1,4 @@
+export { SliderWithValue } from "./SliderWithValue";
+export { PageHeader } from "./PageHeader";
+export { ReactAgentCard } from "./ReactAgentCard";
+export { ContextManagementCard } from "./ContextManagementCard";
