+++ console/src/pages/Chat/index.module.less
@@ -28,6 +28,10 @@
   margin: 0 auto 0 0 !important;
 }
 
+:global(.copaw-chat-anywhere-default-header-right) {
+  margin: 0 auto 0 0 !important;
+}
+
 /* Ensure ModelSelector aligns to left in rightHeader */
 :global([class*="chat-anywhere-default-header-right"]) > *,
 :global([class*="default-header-right"]) > * {
