+++ console/src/pages/Settings/Environments/index.module.less
@@ -1,4 +1,4 @@
-.page {
+.environmentsPage {
   padding: 24px;
 }
