+++ console/src/pages/Settings/Models/index.module.less
@@ -1,4 +1,4 @@
-.page {
+.settingsPage {
   padding: 24px;
 }
