+++ console/src/pages/Settings/Security/components/index.ts
@@ -1,0 +1,4 @@
+export * from "./PageHeader";
+export * from "./RuleTable";
+export * from "./RuleModal";
+export * from "./PreviewModal";
