+++ console/src/pages/Settings/TokenUsage/index.module.less
@@ -1,4 +1,4 @@
-.page {
+.tokenUsagePage {
   padding: 24px;
 }
